+++ packages/generator/src/common/dungeon.c
@@ -2,7 +2,7 @@
 #include <combo/dungeon.h>
 #include <combo/config.h>
 
-void comboDungeonSetFlags(int dungeonId, int flags)
+void Dungeon_SetFlags(int dungeonId, int flags)
 {
     switch (dungeonId)
     {
@@ -13,17 +13,15 @@ void comboDungeonSetFlags(int dungeonId, int flags)
         BITMAP16_SET(gOotSave.info.eventsChk, EV_OOT_CHK_MIDO_TREE_DEAD);
         break;
     case DUNGEONID_TEMPLE_WATER:
+        gMiscFlags.waterBeaten = 1;
         if (flags & DUNGEONCLEARFLAG_EFFECT)
         {
             BITMAP16_SET(gOotSave.info.eventsChk, EV_OOT_CHK_LAKE_HYLIA_WATER);
-            gMiscFlags.wispOotLake = 1;
         }
-        if (flags & DUNGEONCLEARFLAG_WISP)
-            gMiscFlags.wispOotLake = 1;
         break;
     case DUNGEONID_TEMPLE_WOODFALL:
         if (flags & DUNGEONCLEARFLAG_BOSS)
-            gMmExtraBoss.boss |= (1 << 0);
+            gMmExtraBoss.dungeon |= (1 << 0);
         if (flags & DUNGEONCLEARFLAG_EFFECT)
         {
             MM_SET_EVENT_WEEK(EV_MM_WEEK_DUNGEON_WF);
@@ -32,22 +30,18 @@ void comboDungeonSetFlags(int dungeonId, int flags)
                 MM_SET_EVENT_WEEK(EV_MM_WEEK_WOODFALL_TEMPLE_RISE);
             }
         }
-        if (flags & DUNGEONCLEARFLAG_WISP)
-            gMiscFlags.wispMmSwamp = 1;
         break;
     case DUNGEONID_TEMPLE_SNOWHEAD:
         if (flags & DUNGEONCLEARFLAG_BOSS)
-            gMmExtraBoss.boss |= (1 << 1);
+            gMmExtraBoss.dungeon |= (1 << 1);
         if (flags & DUNGEONCLEARFLAG_EFFECT)
         {
             MM_SET_EVENT_WEEK(EV_MM_WEEK_DUNGEON_SH);
         }
-        if (flags & DUNGEONCLEARFLAG_WISP)
-            gMiscFlags.wispMmMountain = 1;
         break;
     case DUNGEONID_TEMPLE_GREAT_BAY:
         if (flags & DUNGEONCLEARFLAG_BOSS)
-            gMmExtraBoss.boss |= (1 << 2);
+            gMmExtraBoss.dungeon |= (1 << 2);
         if (flags & DUNGEONCLEARFLAG_EFFECT)
         {
             MM_SET_EVENT_WEEK(EV_MM_WEEK_DUNGEON_GB);
@@ -56,24 +50,20 @@ void comboDungeonSetFlags(int dungeonId, int flags)
                 MM_SET_EVENT_WEEK(EV_MM_WEEK_GREAT_BAY_TURTLE);
             }
         }
-        if (flags & DUNGEONCLEARFLAG_WISP)
-            gMiscFlags.wispMmOcean = 1;
         break;
     case DUNGEONID_TEMPLE_STONE_TOWER:
     case DUNGEONID_TEMPLE_STONE_TOWER_INVERTED:
         if (flags & DUNGEONCLEARFLAG_BOSS)
-            gMmExtraBoss.boss |= (1 << 3);
+            gMmExtraBoss.dungeon |= (1 << 3);
         if (flags & DUNGEONCLEARFLAG_EFFECT)
         {
             MM_SET_EVENT_WEEK(EV_MM_WEEK_DUNGEON_ST);
         }
-        if (flags & DUNGEONCLEARFLAG_WISP)
-            gMiscFlags.wispMmValley = 1;
         break;
     }
 }
 
-int comboBossDungeon(int dungeonId)
+int Dungeon_Boss(int dungeonId)
 {
     for (int i = 0; i <= BOSSID_TWINMOLD; ++i)
     {
@@ -81,4 +71,73 @@ int comboBossDungeon(int dungeonId)
             return i;
     }
     return -1;
+}
+
+static const u8 kOotBossScenes[] = {
+    SCE_OOT_LAIR_GOHMA,
+    SCE_OOT_LAIR_KING_DODONGO,
+    SCE_OOT_LAIR_BARINADE,
+    SCE_OOT_LAIR_PHANTOM_GANON,
+    SCE_OOT_LAIR_VOLVAGIA,
+    SCE_OOT_LAIR_MORPHA,
+    SCE_OOT_LAIR_BONGO_BONGO,
+    SCE_OOT_LAIR_TWINROVA,
+};
+
+static const u8 kOotBossRooms[] = {
+    1,
+    1,
+    1,
+    1,
+    1,
+    1,
+    1,
+    3,
+};
+
+int DungeonBoss_GetFlag(int bossId)
+{
+    u8 sceneId;
+    u8 roomId;
+
+    if (bossId < BOSSID_ODOLWA)
+    {
+        sceneId = kOotBossScenes[bossId];
+        roomId = kOotBossRooms[bossId];
+
+#if defined(GAME_OOT)
+        if (gPlay->sceneId == sceneId)
+            return Flags_GetClear(gPlay, roomId);
+#endif
+        return gOotSave.info.perm[sceneId].roomClear & (1 << roomId);
+    }
+    else
+    {
+        return gMmExtraBoss.boss & (1 << (bossId - BOSSID_ODOLWA));
+    }
+}
+
+void DungeonBoss_SetFlag(int bossId)
+{
+    u8 sceneId;
+    u8 roomId;
+    u8 flag;
+
+    if (bossId < BOSSID_ODOLWA)
+    {
+        sceneId = kOotBossScenes[bossId];
+        roomId = kOotBossRooms[bossId];
+
+#if defined(GAME_OOT)
+        if (gPlay->sceneId == sceneId)
+            Flags_SetClear(gPlay, roomId);
+#endif
+        gOotSave.info.perm[sceneId].roomClear |= (1 << roomId);
+    }
+    else
+    {
+        flag = (1 << (bossId - BOSSID_ODOLWA));
+        gMmExtraBoss.boss |= flag;
+        gMmExtraBoss.bossCycle |= flag;
+    }
 }