+++ packages/generator/src/oot/save.c
@@ -237,7 +237,6 @@ static void applyStartingEvents(void)
 
 void comboCreateSave(void* unk, void* buffer)
 {
-    int dungeonClearFlags;
     u32 base;
 
     /* Create MM save */
@@ -344,13 +343,10 @@ void comboCreateSave(void* unk, void* buffer)
     applyStartingEvents();
 
     /* Apply pre-completed dungeons */
-    dungeonClearFlags = DUNGEONCLEARFLAG_BOSS;
-    if (Config_Flag(CFG_REGION_STATE_DUNGEONS))
-        dungeonClearFlags |= DUNGEONCLEARFLAG_WISP;
     for (int i = 0; i < 32; ++i)
     {
         if (gComboConfig.preCompleted & (1 << i))
-            comboDungeonSetFlags(i, dungeonClearFlags);
+            Dungeon_SetFlags(i, DUNGEONCLEARFLAG_BOSS);
     }
 
     /* Write save */